+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Update PopupNotificationList:
+    
+    - Closing the notification by clicking on the entire area;
+    
+    - Add new props for notification: `onMouseEnter`, `onMouseLeave`.
+
 ## [0.37.0] - 2021-05-26
 
 - Fixed bug in CopyToClipboard with lost indentations.